backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Task Queue
celery==5.3.6

# NLP & AI
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
numpy==1.26.3
scikit-learn==1.4.0

# News Collection
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
newspaper3k==0.2.8
python-dateutil==2.8.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-slugify==8.0.1
pytz==2023.3
langdetect==1.0.9

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality (dev)
black==24.1.1
flake8==7.0.0
mypy==1.8.0

